Implement `apply role`
