+++ code/cmd/role.go
@@ -2,14 +2,15 @@ package cmd
 
 import (
 	"fmt"
-	"os"
 
 	"github.com/spf13/cobra"
 )
 
+var roleFile string
+
 func init() {
 	gmApplyCmd.AddCommand(gmApplyRoleCmd)
-	gmApplyRoleCmd.Flags().StringVar(&policyFile, "policy", "", "Path to role YAML file or directory (mandatory)")
+	gmApplyRoleCmd.Flags().StringVar(&roleFile, "role", "", "Path to role YAML file or directory (mandatory)")
 	gmApplyRoleCmd.MarkFlagRequired("role")
 }
 
@@ -17,7 +18,12 @@ var gmApplyRoleCmd = &cobra.Command{
 	Use:   "role",
 	Short: "Apply roles from the specified YAML file or directory",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("NYI")
-		os.Exit(1)
+		if err := applyRole(); err != nil {
+			fmt.Println("Error:", err)
+		}
 	},
+}
+
+func applyRole() error {
+	return fmt.Errorf("NYI")
 }
