+++ code/modules/database/base/base.go
@@ -4,7 +4,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/welibekov/grantmaster/modules/policy/types"
+	polTypes "github.com/welibekov/grantmaster/modules/policy/types"
+	rolTypes "github.com/welibekov/grantmaster/modules/role/types"
 )
 
 type Database struct{}
@@ -13,11 +14,11 @@ func NewDatabase() *Database {
 	return &Database{}
 }
 
-func (d *Database) ApplyPolicy(context.Context, []types.Policy) error {
+func (d *Database) ApplyPolicy(context.Context, []polTypes.Policy) error {
 	return d.notImplemented()
 }
 
-func (d *Database) ApplyRole(context.Context, []types.Role) error {
+func (d *Database) ApplyRole(context.Context, []rolTypes.Role) error {
 	return d.notImplemented()
 }
