+++ code/modules/policy/types/types.go
@@ -6,10 +6,3 @@ type Policy struct {
 	Username string   `yaml:"username"` // The username associated with the policy.
 	Roles    []string `yaml:"roles"`    // List of roles allowed to perform the action.
 }
-
-// Role defines a set of roles that should be created in the system.
-type Role struct {
-	Name   string `yaml:"name"`   // The name associated with the role
-	Type   string `yaml:"type"`   // Type of the role.
-	Schema string `yaml:"schema"` // Schema name associated with the role
-}
